% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.form-package.R
\docType{package}
\name{model.form-package}
\alias{model.form}
\alias{model.form-package}
\title{Neutral versioned form definitions}
\description{
Load explicit YAML resources and check their definitions. This package never
validates submitted workbooks, evaluates rules, or supplies reference data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ultreia-io.github.io/model.form/}
  \item \url{https://github.com/ultreia-io/model.form}
  \item Report bugs at \url{https://github.com/ultreia-io/model.form/issues}
}

}
\author{
\strong{Maintainer}: Tony Chemit \email{tchemit@users.noreply.github.com}

Authors:
\itemize{
  \item Tony Chemit \email{tchemit@users.noreply.github.com}
}

}
\keyword{internal}
